% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.survivalsvm-package}
\alias{mlr3learners.survivalsvm}
\alias{mlr3learners.survivalsvm-package}
\title{mlr3learners.survivalsvm: Learners from the {survivalsvm} package for 'mlr3'}
\description{
Adds learners from the {survivalsvm} package to
    'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
